requirements.txt
# Python packages
pyserial
pycryptodome
